+++ server/routes.js
@@ -4,6 +4,8 @@ const router = express.Router();
 const config = require('./config.js');
 var axios = require('axios');
 
+const { getLocationsAndSend } = require('./utils.js');
+
 // APIs
 const gmaps = require('./google-maps.js');
 const yelp = require('./yelp.js');
@@ -78,7 +80,6 @@ var routerInstance = function(io) {
 
             console.log('coordinates1', coordinates1);
             console.log('coordinates2', coordinates2);
-
             // Deprecated
             // get midpoint
             // gmaps.generateMidpoint(coordinates1, coordinates2)
@@ -100,38 +101,7 @@ var routerInstance = function(io) {
             //   });
 
             // send all points
-            gmaps.generatePointsAlong(coordinates1, coordinates2, arrivalTime, transportation)
-              .then(({ pointsAlong, midpoint, departure_time }) => {
-                // Generate midpoint locations with higher search radius
-                console.log('points along', pointsAlong);
-                yelp.yelpRequest(midpoint, 10, 100)
-                  .then((yelpLocations) => {
-                    io.sockets.emit('midpoint', { lat: midpoint.latitude, lng: midpoint.longitude });
-                    io.sockets.emit('mid meeting locations', yelpLocations)
-                    // formatted as { location1: [lat,lng], location2: [lat, lng] }
-                    io.sockets.emit('user locations', {
-                      location1: { lat: coordinates1[0], lng: coordinates1[1] },
-                      location2: { lat: coordinates2[0], lng: coordinates2[1] },
-                    });
-                  });
-                const mappedYelp = pointsAlong.map((point) => {
-                  // points.forEach(point => {
-                  return yelp.yelpRequest(point, 3)
-                    .then((yelpLocations) => {
-                      // Re-render client
-                      return yelpLocations;
-                    });
-                });
-                // Generate all restaurants along the path
-                Promise.all(mappedYelp)
-                  .then((locationsArr) => {
-                    // MERGE ARRAY OF ARRAYS
-                    const allMeetingLocations = [].concat.apply([], locationsArr);
-                    io.sockets.emit('all meeting locations', allMeetingLocations );
-                  })
-                  .catch(err => console.log("Error with promise all"), err);
-              })
-              .catch(err => console.log(err));
+            getLocationsAndSend(coordinates1, coordinates2, arrivalTime, transportation, io);
             res.send('Results found.');
           })
           .catch(err => console.log("Err getting geocode from Google API"), err);
